# Mode 'parallel' for EncSALayer to speed up infer on ONNX