--- a/modules/commons/common_layers.py
+++ b/modules/commons/common_layers.py
@@ -132,18 +132,23 @@ class TransformerFFNLayer(nn.Module):
 
 class EncSALayer(nn.Module):
     def __init__(self, c, num_heads, dropout, attention_dropout=0.1,
-                 relu_dropout=0.1, kernel_size=9, act='gelu'):
+                 relu_dropout=0.1, kernel_size=9, act='gelu', tf_enc_mode='series'):
         super().__init__()
         self.dropout = dropout
         self.layer_norm1 = LayerNorm(c)
         self.self_attn = MultiheadAttention(
             c, num_heads, dropout=attention_dropout, bias=False,
         )
-        self.layer_norm2 = LayerNorm(c)
+        self.tf_enc_mode = tf_enc_mode
+        if tf_enc_mode == 'series':
+            self.layer_norm2 = LayerNorm(c)
+        elif tf_enc_mode == 'parallel':
+            self.layer_norm2 = nn.Identity()
+        else:
+            raise ValueError(f'{tf_enc_mode} is not a valid encoder model type')
         self.ffn = TransformerFFNLayer(
             c, 4 * c, kernel_size=kernel_size, dropout=relu_dropout, act=act
         )
-
     def forward(self, x, encoder_padding_mask=None, **kwargs):
         layer_norm_training = kwargs.get('layer_norm_training', None)
         if layer_norm_training is not None:
@@ -151,20 +156,38 @@ class EncSALayer(nn.Module):
             self.layer_norm2.training = layer_norm_training
         residual = x
         x = self.layer_norm1(x)
-        x, _, = self.self_attn(
-            query=x,
-            key=x,
-            value=x,
-            key_padding_mask=encoder_padding_mask
-        )
-        x = F.dropout(x, self.dropout, training=self.training)
-        x = residual + x
-        x = x * (1 - encoder_padding_mask.float()).transpose(0, 1)[..., None]
+        if self.tf_enc_mode == 'series':
+            x, _, = self.self_attn(
+                query=x,
+                key=x,
+                value=x,
+                key_padding_mask=encoder_padding_mask
+            )
+            x = F.dropout(x, self.dropout, training=self.training)
+            x = residual + x
+            x = x * (1 - encoder_padding_mask.float()).transpose(0, 1)[..., None]
 
-        residual = x
-        x = self.layer_norm2(x)
-        x = self.ffn(x)
-        x = F.dropout(x, self.dropout, training=self.training)
-        x = residual + x
-        x = x * (1 - encoder_padding_mask.float()).transpose(0, 1)[..., None]
+            residual = x
+            x = self.layer_norm2(x)
+            x = self.ffn(x)
+            x = F.dropout(x, self.dropout, training=self.training)
+            x = residual + x
+            x = x * (1 - encoder_padding_mask.float()).transpose(0, 1)[..., None]
+        elif self.tf_enc_mode == 'parallel':
+            # transformer-parallel from GPT-J
+            x_attn, _, = self.self_attn(
+                query=x,
+                key=x,
+                value=x,
+                key_padding_mask=encoder_padding_mask
+            )
+            x_attn = F.dropout(x_attn, self.dropout, training=self.training)
+            x_ffn = self.ffn(x)
+            x_ffn = F.dropout(x_ffn, self.dropout, training=self.training)
+            
+            x = (x_attn + x_ffn) + residual
+            x = x * (1 - encoder_padding_mask.float()).transpose(0, 1)[..., None]
+        else:
+           raise ValueError(f'{tf_enc_mode} is not a valid encoder model type')
+
         return x